+++ src/internal/m365/exchange/events_restore.go
@@ -1,6 +1,7 @@
 package exchange
 
 import (
+	"bytes"
 	"context"
 	"fmt"
 	"strings"
@@ -14,7 +15,9 @@ import (
 	"github.com/alcionai/corso/src/internal/common/str"
 	"github.com/alcionai/corso/src/internal/m365/graph"
 	"github.com/alcionai/corso/src/pkg/backup/details"
+	"github.com/alcionai/corso/src/pkg/control"
 	"github.com/alcionai/corso/src/pkg/fault"
+	"github.com/alcionai/corso/src/pkg/logger"
 	"github.com/alcionai/corso/src/pkg/path"
 	"github.com/alcionai/corso/src/pkg/services/m365/api"
 )
@@ -23,17 +26,13 @@ var _ itemRestorer = &eventRestoreHandler{}
 
 type eventRestoreHandler struct {
 	ac api.Events
-	ip itemPoster[models.Eventable]
 }
 
 func newEventRestoreHandler(
 	ac api.Client,
 ) eventRestoreHandler {
-	ace := ac.Events()
-
 	return eventRestoreHandler{
-		ac: ace,
-		ip: ace,
+		ac: ac.Events(),
 	}
 }
 
@@ -72,6 +71,32 @@ func (h eventRestoreHandler) restore(
 	ctx context.Context,
 	body []byte,
 	userID, destinationID string,
+	collisionKeyToItemID map[string]string,
+	collisionPolicy control.CollisionPolicy,
+	errs *fault.Bus,
+) (*details.ExchangeInfo, error) {
+	return restoreEvent(
+		ctx,
+		h.ac,
+		body,
+		userID, destinationID,
+		collisionKeyToItemID,
+		collisionPolicy,
+		errs)
+}
+
+type eventRestorer interface {
+	postItemer[models.Eventable]
+	eventInstanceAndAttachmenter
+}
+
+func restoreEvent(
+	ctx context.Context,
+	er eventRestorer,
+	body []byte,
+	userID, destinationID string,
+	collisionKeyToItemID map[string]string,
+	collisionPolicy control.CollisionPolicy,
 	errs *fault.Bus,
 ) (*details.ExchangeInfo, error) {
 	event, err := api.BytesToEventable(body)
@@ -80,6 +105,18 @@ func (h eventRestoreHandler) restore(
 	}
 
 	ctx = clues.Add(ctx, "item_id", ptr.Val(event.GetId()))
+	collisionKey := api.EventCollisionKey(event)
+
+	if _, ok := collisionKeyToItemID[collisionKey]; ok {
+		log := logger.Ctx(ctx).With("collision_key", clues.Hide(collisionKey))
+		log.Debug("item collision")
+
+		// TODO(rkeepers): Replace probably shouldn't no-op.  Just a starting point.
+		if collisionPolicy == control.Skip || collisionPolicy == control.Replace {
+			log.Debug("skipping item with collision")
+			return nil, graph.ErrItemAlreadyExistsConflict
+		}
+	}
 
 	event = toEventSimplified(event)
 
@@ -92,14 +129,14 @@ func (h eventRestoreHandler) restore(
 		event.SetAttachments(nil)
 	}
 
-	item, err := h.ip.PostItem(ctx, userID, destinationID, event)
+	item, err := er.PostItem(ctx, userID, destinationID, event)
 	if err != nil {
 		return nil, graph.Wrap(ctx, err, "restoring calendar item")
 	}
 
 	err = uploadAttachments(
 		ctx,
-		h.ac,
+		er,
 		attachments,
 		userID,
 		destinationID,
@@ -117,7 +154,15 @@ func (h eventRestoreHandler) restore(
 	}
 
 	// Fix up event instances in case we have a recurring event
-	err = updateRecurringEvents(ctx, h.ac, userID, destinationID, ptr.Val(item.GetId()), event)
+	err = updateRecurringEvents(
+		ctx,
+		er,
+		userID,
+		destinationID,
+		ptr.Val(item.GetId()),
+		event,
+		errs,
+	)
 	if err != nil {
 		return nil, clues.Stack(err)
 	}
@@ -130,9 +175,10 @@ func (h eventRestoreHandler) restore(
 
 func updateRecurringEvents(
 	ctx context.Context,
-	ac api.Events,
+	eiaa eventInstanceAndAttachmenter,
 	userID, containerID, itemID string,
 	event models.Eventable,
+	errs *fault.Bus,
 ) error {
 	if event.GetRecurrence() == nil {
 		return nil
@@ -144,12 +190,12 @@ func updateRecurringEvents(
 	cancelledOccurrences := event.GetAdditionalData()["cancelledOccurrences"]
 	exceptionOccurrences := event.GetAdditionalData()["exceptionOccurrences"]
 
-	err := updateCancelledOccurrences(ctx, ac, userID, itemID, cancelledOccurrences)
+	err := updateCancelledOccurrences(ctx, eiaa, userID, itemID, cancelledOccurrences)
 	if err != nil {
 		return clues.Wrap(err, "update cancelled occurrences")
 	}
 
-	err = updateExceptionOccurrences(ctx, ac, userID, itemID, exceptionOccurrences)
+	err = updateExceptionOccurrences(ctx, eiaa, userID, containerID, itemID, exceptionOccurrences, errs)
 	if err != nil {
 		return clues.Wrap(err, "update exception occurrences")
 	}
@@ -157,15 +203,35 @@ func updateRecurringEvents(
 	return nil
 }
 
+type eventInstanceAndAttachmenter interface {
+	attachmentGetDeletePoster
+	DeleteItem(
+		ctx context.Context,
+		userID, itemID string,
+	) error
+	GetItemInstances(
+		ctx context.Context,
+		userID, itemID string,
+		startDate, endDate string,
+	) ([]models.Eventable, error)
+	PatchItem(
+		ctx context.Context,
+		userID, eventID string,
+		body models.Eventable,
+	) (models.Eventable, error)
+}
+
 // updateExceptionOccurrences take events that have exceptions, uses
 // the originalStart date to find the instance and modify it to match
 // the backup by updating the instance to match the backed up one
 func updateExceptionOccurrences(
 	ctx context.Context,
-	ac api.Events,
+	eiaa eventInstanceAndAttachmenter,
 	userID string,
+	containerID string,
 	itemID string,
 	exceptionOccurrences any,
+	errs *fault.Bus,
 ) error {
 	if exceptionOccurrences == nil {
 		return nil
@@ -193,9 +259,11 @@ func updateExceptionOccurrences(
 		startStr := dttm.FormatTo(start, dttm.DateOnly)
 		endStr := dttm.FormatTo(start.Add(24*time.Hour), dttm.DateOnly)
 
+		ictx := clues.Add(ctx, "event_instance_id", ptr.Val(evt.GetId()), "event_instance_date", start)
+
 		// Get all instances on the day of the instance which should
 		// just the one we need to modify
-		evts, err := ac.GetItemInstances(ctx, userID, itemID, startStr, endStr)
+		instances, err := eiaa.GetItemInstances(ictx, userID, itemID, startStr, endStr)
 		if err != nil {
 			return clues.Wrap(err, "getting instances")
 		}
@@ -203,25 +271,160 @@ func updateExceptionOccurrences(
 		// Since the min recurrence interval is 1 day and we are
 		// querying for only a single day worth of instances, we
 		// should not have more than one instance here.
-		if len(evts) != 1 {
+		if len(instances) != 1 {
 			return clues.New("invalid number of instances for modified").
-				With("instances_count", len(evts), "search_start", startStr, "search_end", endStr)
+				With("instances_count", len(instances), "search_start", startStr, "search_end", endStr)
 		}
 
 		evt = toEventSimplified(evt)
 
-		// TODO(meain): Update attachments (might have to diff the
-		// attachments using ids and delete or add). We will have
-		// to get the id of the existing attachments, diff them
-		// with what we need a then create/delete items kinda like
-		// permissions
-		_, err = ac.PatchItem(ctx, userID, ptr.Val(evts[0].GetId()), evt)
+		_, err = eiaa.PatchItem(ictx, userID, ptr.Val(instances[0].GetId()), evt)
 		if err != nil {
 			return clues.Wrap(err, "updating event instance")
+		}
+
+		// We are creating event again from map as `toEventSimplified`
+		// removed the attachments and creating a clone from start of
+		// the event is non-trivial
+		evt, err = api.EventFromMap(instance)
+		if err != nil {
+			return clues.Wrap(err, "parsing event instance")
+		}
+
+		err = updateAttachments(
+			ictx,
+			eiaa,
+			userID,
+			containerID,
+			ptr.Val(instances[0].GetId()),
+			evt,
+			errs)
+		if err != nil {
+			return clues.Wrap(err, "updating event instance attachments")
 		}
 	}
 
 	return nil
+}
+
+type attachmentGetDeletePoster interface {
+	attachmentPoster
+	GetAttachments(
+		ctx context.Context,
+		immutableIDs bool,
+		userID string,
+		itemID string,
+	) ([]models.Attachmentable, error)
+	DeleteAttachment(
+		ctx context.Context,
+		userID, calendarID, eventID, attachmentID string,
+	) error
+}
+
+// updateAttachments updates the attachments of an event to match what
+// is present in the backed up event. Ideally we could make use of the
+// id of the series master event's attachments to see if we had
+// added/removed any attachments, but as soon an event is modified,
+// the id changes which makes the ids unusable. In this function, we
+// use the name and content bytes to detect the changes. This function
+// can be used to update the attachments of any event irrespective of
+// whether they are event instances of a series master although for
+// newer event, since we probably won't already have any events it
+// would be better use Post[Small|Large]Attachment.
+func updateAttachments(
+	ctx context.Context,
+	agdp attachmentGetDeletePoster,
+	userID, containerID, eventID string,
+	event models.Eventable,
+	errs *fault.Bus,
+) error {
+	el := errs.Local()
+
+	attachments, err := agdp.GetAttachments(ctx, false, userID, eventID)
+	if err != nil {
+		return clues.Wrap(err, "getting attachments")
+	}
+
+	// Delete attachments that are not present in the backup but are
+	// present in the event(ones that were automatically inherited
+	// from series master).
+	for _, att := range attachments {
+		if el.Failure() != nil {
+			return el.Failure()
+		}
+
+		name := ptr.Val(att.GetName())
+		id := ptr.Val(att.GetId())
+
+		content, err := api.GetAttachmentContent(att)
+		if err != nil {
+			return clues.Wrap(err, "getting attachment").With("attachment_id", id)
+		}
+
+		found := false
+
+		for _, nAtt := range event.GetAttachments() {
+			nName := ptr.Val(nAtt.GetName())
+
+			nContent, err := api.GetAttachmentContent(nAtt)
+			if err != nil {
+				return clues.Wrap(err, "getting attachment").With("attachment_id", ptr.Val(nAtt.GetId()))
+			}
+
+			if name == nName && bytes.Equal(content, nContent) {
+				found = true
+				break
+			}
+		}
+
+		if !found {
+			err = agdp.DeleteAttachment(ctx, userID, containerID, eventID, id)
+			if err != nil {
+				logger.CtxErr(ctx, err).With("attachment_name", name).Info("attachment delete failed")
+				el.AddRecoverable(ctx, clues.Wrap(err, "deleting event attachment").
+					WithClues(ctx).With("attachment_name", name))
+			}
+		}
+	}
+
+	// Upload missing(attachments that are present in the individual
+	// instance but not in the series master event) attachments
+	for _, att := range event.GetAttachments() {
+		name := ptr.Val(att.GetName())
+		id := ptr.Val(att.GetId())
+
+		content, err := api.GetAttachmentContent(att)
+		if err != nil {
+			return clues.Wrap(err, "getting attachment").With("attachment_id", id)
+		}
+
+		found := false
+
+		for _, nAtt := range attachments {
+			nName := ptr.Val(nAtt.GetName())
+
+			bContent, err := api.GetAttachmentContent(nAtt)
+			if err != nil {
+				return clues.Wrap(err, "getting attachment").With("attachment_id", ptr.Val(nAtt.GetId()))
+			}
+
+			// Max size allowed for an outlook attachment is 150MB
+			if name == nName && bytes.Equal(content, bContent) {
+				found = true
+				break
+			}
+		}
+
+		if !found {
+			err = uploadAttachment(ctx, agdp, userID, containerID, eventID, att)
+			if err != nil {
+				return clues.Wrap(err, "uploading attachment").
+					With("attachment_id", id)
+			}
+		}
+	}
+
+	return el.Failure()
 }
 
 // updateCancelledOccurrences get the cancelled occurrences which is a
@@ -229,7 +432,7 @@ func updateExceptionOccurrences(
 // that and uses the to get the event instance at that date to delete.
 func updateCancelledOccurrences(
 	ctx context.Context,
-	ac api.Events,
+	eiaa eventInstanceAndAttachmenter,
 	userID string,
 	itemID string,
 	cancelledOccurrences any,
@@ -266,7 +469,7 @@ func updateCancelledOccurrences(
 
 		// Get all instances on the day of the instance which should
 		// just the one we need to modify
-		evts, err := ac.GetItemInstances(ctx, userID, itemID, startStr, endStr)
+		instances, err := eiaa.GetItemInstances(ctx, userID, itemID, startStr, endStr)
 		if err != nil {
 			return clues.Wrap(err, "getting instances")
 		}
@@ -274,16 +477,28 @@ func updateCancelledOccurrences(
 		// Since the min recurrence interval is 1 day and we are
 		// querying for only a single day worth of instances, we
 		// should not have more than one instance here.
-		if len(evts) != 1 {
+		if len(instances) != 1 {
 			return clues.New("invalid number of instances for cancelled").
-				With("instances_count", len(evts), "search_start", startStr, "search_end", endStr)
+				With("instances_count", len(instances), "search_start", startStr, "search_end", endStr)
 		}
 
-		err = ac.DeleteItem(ctx, userID, ptr.Val(evts[0].GetId()))
+		err = eiaa.DeleteItem(ctx, userID, ptr.Val(instances[0].GetId()))
 		if err != nil {
 			return clues.Wrap(err, "deleting event instance")
 		}
 	}
 
 	return nil
+}
+
+func (h eventRestoreHandler) getItemsInContainerByCollisionKey(
+	ctx context.Context,
+	userID, containerID string,
+) (map[string]string, error) {
+	m, err := h.ac.GetItemsInContainerByCollisionKey(ctx, userID, containerID)
+	if err != nil {
+		return nil, err
+	}
+
+	return m, nil
 }