+++ CHANGELOG.md
@@ -20,12 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix Exchange backup issue caused by incorrect json serialization
 - Fix issues with details model containing duplicate entry for api consumers
 - Handle OLE conversion errors when trying to fetch attachments
+- Fix uploading large attachments for emails and calendar
 
 ### Changed
 - Do not display all the items that we restored at the end if there are more than 15. You can override this with `--verbose`.
-
-### Known Issues
-- Changes to attachments in instances of recurring events compared to the series master aren't restored
 
 ## [v0.8.0] (beta) - 2023-05-15